move checkProfiles()
function & tests were in the wrong location
sso/cache => sso/roles
