+++ sso/cache.go
@@ -522,28 +522,3 @@ func (c *Cache) addConfigRoles(r *Roles, config *SSOConfig) error {
 	}
 	return nil
 }
-
-// checkProfiles verfies that all the Profile names are unique for all the defined roles
-func (r *Roles) checkProfiles(s *Settings) error {
-	profileUniqueCheck := map[string]string{} // ProfileName() => Arn
-	for accountId, account := range r.Accounts {
-		for roleName, role := range account.Roles {
-			flat, err := r.GetRole(accountId, roleName)
-			if err != nil {
-				return err
-			}
-
-			pname, err := flat.ProfileName(s)
-			if err != nil {
-				return err
-			}
-
-			if arn, duplicate := profileUniqueCheck[pname]; duplicate {
-				return fmt.Errorf("Duplicate profile name '%s' for:\n- %s\n- %s", pname, arn, role.Arn)
-			} else {
-				profileUniqueCheck[pname] = arn
-			}
-		}
-	}
-	return nil
-}
